+++ CHANGELOG.md
@@ -28,6 +28,12 @@
 
   Migration: call `new ClientId(id)` to construct a `ClientId`, and `id.as_bytes()` to get a byte array out.
 
+- `ClientId` wrapper accepts a byte array, not a string.
+
+  Affected platforms: Kotlin
+
+  Migration: call `.toByteArray()` on the input.
+
 ## v7.0.1 - 2025-06-02
 
 ### Bug Fixes